Removed Mouse, Updated Documentation.
Removed the dependency on Mouse.  This will need to be pushed along with
AnyEvent::Blackboard being updated which hasn't been done yet.

Reordered the methods so they appear in a more natural fashion in the
documentation, and update the documentation to have a better description of the
component.

Created a new accessor ``hungup'' which describes, as one might hopefully
guess, whether or not hangup has been called.  Made the tests depend on this,
rather than the private value ``_hangup'' which is now gone.  This also means
that was the only real API encapsulation violation in the unit tests.